doc: fix passing non-event messages in distill examples
The distill filter example means to let non-event messages go through,
but it actually drops them.  Fix it so it actually lets them through.

Verified by sending the output of the filter to `sink.text.details`.
Before the patch, there is no stream beginning message.  After the
patch, there is.

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: Ib463a2e931c61d481430f3f7120a251d66f23bc3