fix: tests: use a portable invocation of 'sed -i'
The GNU and BSD sed implementations slightly differ in the handling of
an argument to the '-i' switch to use as a suffix to a backup file. BSD
sed requires an argument while it's optional for GNU sed, if present it
must be placed directly after the switch with no space.

To make both implementations work properly and not create a backup file,
a zero length argument must be placed directly after the switch with no
space.

Signed-off-by: Michael Jeanson <mjeanson@efficios.com>
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: Ie1584fbdaf9d8752190f10288b2a1a89e3a43ad1