tests: refactor auto discovery log-level/params tests
We have tests (both CLI and Python) that verify the parameters, log
levels and Python objects passed to the components instantiated by the
automatic source discovery mechanism.  In order for the tests to be able
to see what the instantiated components received, the components create
some streams and encode the requested information in the stream name.
The CLI test retrieves it via the details sink and the Python test
retrieves it via the Python API.

Refactor this to be more straightforward:

 - When testing the Python API, pass some objects down (available as the
   `obj` init parameter) where the instantiated components put the
   requested information.
 - When testing from the CLI (in which case `obj` is `None`), print the
   requested information on stdout.

In the CLI tests, we need to sort the output, because the order in which
the lines are printed is not guaranteed (it depends the order
autodisc/glib walks the filesystem).  We previously didn't have this
problem, because we were relying on stream beginning messages, which
were sorted in a stable way by the muxer.

I initially wrote this change because I wanted to use it to
test the log levels passed to support-info queries during source
auto-discovery.  I later realized it would not be possible to pass down
Python objects through the auto discovery all the way to support-info
queries, so it won't be possible to use the same pattern.  However, I
think this change still simplifies how things work.

Change-Id: I6daa2c38b13832a23ac1446754fe5a7102a7b645
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/11591
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>