configure: use single word for PROJECT-NAME argument of DX_INIT_DOXYGEN
Running configure with --enable-api-doc shows this warning:

    /home/simark/src/babeltrace/configure: line 25508: 2: command not found

The line 25508 in question is:

  DX_PROJECT=Babeltrace 2

In shell script, this tries to execute the command `2`, with the
`DX_PROJECT` environment variable set to `Babeltrace`.  This is
obviously not what we want.

The documentation for the DX_INIT_DOXYGEN macro says:

   ... call DX_INIT_DOXYGEN with the following parameters: a one-word
   name for the project for use as a filename base etc.

So, we shouldn't pass something with a space as the project name to
DX_INIT_DOXYGEN.  Change it to `babeltrace2`.  I don't think that this
gets used anywhere though, as I didn't see a difference in my build
directory before and after the change (except in the Makefiles, which
don't use the value themselves).

Reported-by: Brice Videau <bvideau@anl.gov>
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I711951f950f2edd99a2ba250ef66dfd9bcc3be68