cpp-common/bt2: add `ConstError`
I would like to re-implement `format_bt_error()` in C++, keeping the C
API as a front-end.  To do so, it would be useful to have a `ConstError`
type, to wrap the `const bt_error *` received as a parameter (we can't
and don't want to use `UniqueConstError` for that).

Add a new `ConstError` class and extract the functionality common with
`UniqueConstError` to a mixin class.

Change-Id: I70f1ddfca56727c05476b0887c838ceb2c414af4
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/13422
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>