pre-commit autoupdate
Run `pre-commit autoupdate` to bump the versions of the
linting/formatting tools.

The black version used to be limited by the Debian or Ubuntu version on
the CI.  Now that the workers run the bleeding edge Debian 13, let's try
to go to the latest version.

flake8 complains with warnings such as:

    src/bindings/python/bt2/bt2/graph.py:61:5: E704 multiple statements on one line (def)

It doesn't like the "..." that black now puts on the same line as the
function signature.  The formatting is dictated by black, so ignore that
warning.

Change-Id: I7e76a0dcb7fe4f17e66911bb558dbf0610608122
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/15244
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>
Tested-by: jenkins <jenkins@lttng.org>