tests/bindings: fix PackageTestCase always testing the same attribute
The way this test is written is not good.  The function `test_func`,
defined inside the loop, does not capture the current value of `name`.
All test methods therefore test with the same `name` value.  I confirmed
this with a print inside `PackageTestCase._assert_in_bt2`.

Change it to have an external function that returns the function to use
as the test method.  This way, each test method will refer to its own
`name`.

I found this because I happened to have "flake8-bugbear" installed
locally, and flake8 showed this warning:

    ./test_package.py:309:29: B023 Function definition does not bind loop variable 'name'.

Change-Id: I97a1a417801f55a822279068306f4d1ccdc8c801
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Reviewed-on: https://review.lttng.org/c/babeltrace/+/15247
Tested-by: jenkins <jenkins@lttng.org>
Reviewed-by: Philippe Proulx <eeppeliteloop@gmail.com>