lib: add `LIBBABELTRACE2_FORCE_COMPONENT_LOG_LEVEL` support
The new `LIBBABELTRACE2_FORCE_COMPONENT_LOG_LEVEL` environment variable
makes it possible to force the log level of all or specific components
added to any trace processing graph.

Its value is one of:

• A single log level value.

  In this case, libbabeltrace2 overrides the log level specified when
  adding _any_ component to any trace processing graph.

• A semicolon-delimited list of `PAT:LEVEL` specifiers.

  `PAT` is a globbing pattern where only the wildcard character, `*`,
  is special: it matches zero or more characters.

  `LEVEL` is a log level value.

  In this case, libbabeltrace2 overrides the log level specified when
  adding a component of which the name matches `PAT` to any trace
  processing graph.

The accepted log level values are the same as
with `LIBBABELTRACE2_INIT_LOG_LEVEL`.

libbabeltrace2 dynamically checks this environment variable before
adding a component to a trace processing graph, meaning you can modify
it at any time with setenv(3).

Example:

    *ctf*:I;my-sink:D;*:W

This is useful for quick debugging when you don't control the component
additions, for example when using `bt2.TraceCollectionMessageIterator`
with an automatically discovered component. I added it at the
libbabeltrace2 level to make this available to any component addition.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I4de174e57d747450ac379071aa25ef14fbec4b51
Reviewed-on: https://review.lttng.org/c/babeltrace/+/15674