configure.ac: require Python bindings for tests
As we're diving into the conversion to a pytest-oriented test suite,
we'll need less of the CLI and more of the bindings. This will make it
possible to stay into the Python realm and avoid useless subprocesses.

Because Python is already required for the tests, I think it's not so
bad to require our bindings too. SWIG 3, released in 2014, is available
everywhere at this point.

Make the Python bindings enabled by default so that you don't need to
pass `--enable-python-bindings` without `--disable-tests`
(sane default).

This also means that, because `--enable-python-bindings` requires
`--enable-shared` currently, tests require shared libraries. The only
way to allow `--disable-shared` is to also pass `--disable-tests`,
`--disable-python-bindings`, and `--enable-built-in-plugins`.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I5c04159ccb0d05c45e67b910ccf82f9733c08c74
Reviewed-on: https://review.lttng.org/c/babeltrace/+/15689
Tested-by: jenkins <jenkins@lttng.org>