tests/conftest.py: add support for compiled Catch2 tests
This patch adds custom pytest collection of Catch2 tests.

The custom collection considers any `test-*.cpp` to be the main source
of a corresponding compiled test in the build directory. For example,
`src/tests/lib/test-plugins.cpp` is the source
of `build/tests/lib/test-plugins`.

Using the source file as the test file instead of the compiled file will
make it possible to run pytest once in the `tests` directory of the
source tree and "see" all the tests.

This means you can run a specific test like this:

    $ bt-pytest tests/lib/test-something.cpp::test_something_works

The pytest file object runs `test-something --list-test-names-only` to
obtain a flat list of test names. In Catch2, those test names can
contain any character: we convert the name to lowercase, then spaces,
`:`, and `-` to underscores, then remove anything that's not
alphanumeric or an underscore, and then prepend `test_`. For example,
the test `Source component class: description is expected`
becomes `test_source_component_class_description_is_expected`.

On test failure, the report includes the full standard output and
error streams.

The pytest file object skips the file entirely if the corresponding
binary doesn't exist, therefore honouring conditions at the
`Makefile.am` level.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: Icc8181591d32c3e96412fd6512c5b077bfc68b8a