tests/utils/python: move external project sources to new `vendor` dir.
This organization is similar to `src/cpp-common/vendor`.

As we add external projects to `tests/utils/python`, it's easier to
deal with them when they're all in the `vendor` directory.

For example, excluding all of them from our linting configuration
becomes straightforward and it won't need to change if we add more.

`tests/utils/utils.sh` changed to add `tests/utils/python/vendor` to
`PYTHONPATH` too.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I8f57e031f8cf217b666a041098bc7a2f9e183fba
Reviewed-on: https://review.lttng.org/c/babeltrace/+/15702
Tested-by: jenkins <jenkins@lttng.org>