tests: bt_tests_cli_utils.py: add run_cli()
This new function runs the built `babeltrace2-for-tests` CLI with custom
arguments and optional extra plugin paths.

The `args` list can contain both strings (passed as is) or instances of
`CliParams`. A `CliParams` object contains a dictionary of structured
parameters. run_cli() converts this dictionary to one or more
`--params` options.

This function uses run() so that libbabeltrace2 can find the built `bt2`
package to load and execute Python plugins if needed.

The `check`, `text`, and `timeout` parameters are forwarded
to btu.run().

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I284f5138ad14371f229064b9b15395e3be29dccb