sink.ctf.fs: add support to create LTTng index
The LTTng index provides data to efficiently seek in an CTF file by
denoting the sizes and offsets of the packets. This helps when working
with large traces that are only partially read (like in trace-compass).
The source.ctf.fs component is already capable of reading this index,
however the corresponding sink component lacks support for writing it.

We add support to generate the index file, if explicitly requested by
the user.

Philippe changes:

• Switched index I/O from `bt_ctfser` API to plain `FILE *`.

• Fixed broken include path.

• Fixed error path bug that leaked a half-built stream.

• Removed dead/no-op code (fs_sink_stream_open_packet() offset update,
  unused `content_size`).

• Made index helpers static and renamed for clarity.

• mkdir() → g_mkdir() for portability.

• Cleaner error messages and comments.

• The component automatically creates an LTTng index when the
  corresponding trace IR trace has an environment which suggests
  an LTTng tracer.

  It also supports never writing an LTTng index or always writing one.

Refs: GH-123
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I716357b08a6f5ca45975b2922e95586e2a817268
Reviewed-on: https://review.lttng.org/c/babeltrace/+/15766
Reviewed-by: Simon Marchi <simon.marchi@efficios.com>
Tested-by: jenkins <jenkins@lttng.org>