tests: convert `test-bitfield.c` to Catch2 + run with pytest
I did the minimal effort for the conversion here: there are seven basic
test cases and no assertions on the success path.

I also expanded the loop of ten random value test cases so that we can
parallelize them easily: they're named `Random values (0)`, `Random
values (1)`, and so on.

Everything else is as is.

The reason is that we don't care so much about the granularity of test
cases for this since it tests a header which hasn't moved in 15 years
and will stay as is forever.

Keeping the `check-no-bitfield` target in `tests/Makefile.am` which sets
the `BT_TESTS_NO_BITFIELD` environment variable to `1` before running
`pytest-tests.sh`. The latter doesn't collect the `tests/bitfield`
location in that case.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I5c603a4a5f12850399beac938ad0fddb40ee9dc1