sink.ctf.fs: add special handling for cpu_id on lttng
When creating an LTTng index, we must not escape the package.context
cpu_id field, as this is part of the protocol between LTTng and
trace-compass (only when using an index). If this field is escaped,
trace-compass is not able to read the created trace.

In the light of the upcoming CTF 2 which does not need the complicated
escaping logic, we just add special handling for this known case.
For details, also see [1].

Philippe: changing it to refer to a known Trace Compass bug because this
is actually why this patch exists.

[1] https://lists.lttng.org/pipermail/lttng-dev/2025-December/031035.html

Refs: GH-123
Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I296b95d22093fd406cd20686b5b77bf24d337dba
Reviewed-on: https://review.lttng.org/c/babeltrace/+/15890
Tested-by: jenkins <jenkins@lttng.org>